pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "octotiagoocta"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "langchain (>=0.3.20,<0.4.0)",
    "langchain-google-genai (>=2.0.11,<3.0.0)",
    "pygithub (>=2.6.1,<3.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "supabase (>=2.13.0,<3.0.0)",
    "google-genai (>=1.5.0,<2.0.0)",
    "colorama (>=0.4.6,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
